% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{adjust_p}
\alias{adjust_p}
\title{Given a set of p-values, returns p-values adjusted using one of several methods.}
\usage{
adjust_p(pvalues, method = "bonferroni")
}
\arguments{
\item{pvalues}{numeric vector of p-values (possibly with NAs). Any other R object is coerced by as.numeric.}

\item{method}{correction method, a character string. Can be BH, BY, bonferroni, holmmel, hochberg, or none.}
}
\value{
numeric vector of corrected p-values.
}
\description{
Given a set of p-values, returns p-values adjusted using one of several methods.
}
