% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{analyze_degs}
\alias{analyze_degs}
\title{Compute the DEGs}
\usage{
analyze_degs(
  expr_mat,
  groups,
  pCutoff = 0.05,
  log2FCCutoff = 1,
  method = "t.test",
  suffix = "_ttest",
  folder = ".",
  min.count = 0,
  padjust_method = "hochberg",
  species = "Mus.musculus"
)
}
\arguments{
\item{expr_mat}{a data frame which columns are genes, rows are samples. the values are counts, tpm or fpkm.}

\item{groups}{a factor which contains group information.}

\item{pCutoff}{a cutoff for p.value.}

\item{log2FCCutoff}{a cutoff for log2FC.}

\item{method}{one of the t.test, wilcox.test, limma.}

\item{min.count}{numeric. Minimum count required for at least some samples (when counts).}
}
\value{
a data frame which contains gene, P.Value, statistic, logFC and AveExpr columns.
}
\description{
Compute the DEGs
}
