% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{anno_genes}
\alias{anno_genes}
\title{Annotate gene ids.}
\usage{
anno_genes(genes, species = "Mus.musculus")
}
\arguments{
\item{genes}{string vector of gene id.}

\item{species}{Homo.sapiens or Mus.musculus, default is Mus.musculus.}
}
\value{
a data frame which contains ensembl_id, entrez_id and gene_symbol columns.
}
\description{
Annotate gene ids.
}
