% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot}
\alias{boxplot}
\title{Draw boxplot}
\usage{
boxplot(d, method = "t.test", jitter_size = 0.4, log_scale = FALSE)
}
\arguments{
\item{d}{a dataframe which contains these columns (gene_symbol, group, organ, value).}

\item{method}{("t.test", "wilcox.test", "anova", "kruskal.test")}
}
\value{
ggplot object
}
\description{
This function need a dataframe which contains three columns (gene_symbol, group, organ, value).
The organ column is optional and the gene symbol column maybe have several genes. The value column
has mean values of the related gene's expression value.
More details on
\enumerate{
\item Points: https://r-graph-gallery.com/89-box-and-scatter-plot-with-ggplot2.html
\item Grouped Boxplot: https://r-graph-gallery.com/265-grouped-boxplot-with-ggplot2.html
\item P-Value: http://www.sthda.com/english/articles/24-ggpubr-publication-ready-plots/76-add-p-values-and-significance-levels-to-ggplots/
}
}
