% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts2fpkm.R
\name{counts2fpkm}
\alias{counts2fpkm}
\title{Convert counts to FPKM}
\usage{
counts2fpkm(counts)
}
\arguments{
\item{counts}{A data frame which rownames are ensembl id and colnames are sample name.}
}
\value{
a data frame which contains FPKM values.
}
\description{
Convert counts to FPKM
}
