% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts2tpm.R
\name{counts2tpm}
\alias{counts2tpm}
\title{Convert counts to TPM}
\usage{
counts2tpm(counts)
}
\arguments{
\item{counts}{A data frame which rownames are ensembl id and colnames are sample name.}
}
\value{
a data frame which contains TPM values.
}
\description{
Convert counts to TPM
}
