% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_charts.R
\name{draw_barplot}
\alias{draw_barplot}
\title{Draw barplot}
\usage{
draw_barplot(deg, ptitle)
}
\arguments{
\item{deg}{a data frame}

\item{ptitle}{plot title}
}
\value{
ggplot object
}
\description{
Draw barplot
}
