% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_charts.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Draw heatmap}
\usage{
draw_heatmap(exp_mat, removed_samples = c(), genes = c(), scale = "row")
}
\arguments{
\item{removed_samples}{Which samples to remove}

\item{genes}{selected genes}

\item{expMatrix}{matrix}
}
\value{
ggplot2 object
}
\description{
Draw heatmap
}
