% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_charts.R
\name{draw_pca}
\alias{draw_pca}
\title{Draw pca plot}
\usage{
draw_pca(expMatrix, is_qc = FALSE, scale = FALSE)
}
\arguments{
\item{expMatrix}{matrix}

\item{is_qc}{boolean}

\item{scale}{a logical value indicating whether the variables should be scaled to have unit variance before the analysis takes place. The default is FALSE for consistency with S, but in general scaling is advisable. Alternatively, a vector of length equal the number of columns of x can be supplied.}
}
\value{
ggplot2 object
}
\description{
Draw pca plot
}
