% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_charts.R
\name{draw_volcano}
\alias{draw_volcano}
\title{Draw volcano plot}
\usage{
draw_volcano(deg, title, pCutoff = 0.05, FCcutoff = 1)
}
\arguments{
\item{deg}{a data frame}

\item{title}{plot title}

\item{pCutoff}{cut-off for statistical significance. A horizontal line will be drawn at -log10(pCutoff).}

\item{FCcutoff}{cut-off for absolute log2 fold-change. Vertical lines will be drawn at the negative and positive values of log2FCcutoff.}
}
\value{
ggplot object
}
\description{
Draw volcano plot
}
