% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_charts.R
\name{draw_volcano_ggplot}
\alias{draw_volcano_ggplot}
\title{Draw volcana with ggplot2}
\usage{
draw_volcano_ggplot(deg, title, pCutoff = 0.05, FCcutoff = 1)
}
\arguments{
\item{deg}{degs}

\item{title}{plot title}

\item{pCutoff}{cut-off for statistical significance. A horizontal line will be drawn at -log10(pCutoff).}

\item{FCcutoff}{cut-off for absolute log2 fold-change. Vertical lines will be drawn at the negative and positive values of log2FCcutoff.}
}
\value{
ggplot2 object
}
\description{
Draw volcana with ggplot2
}
