% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{enricher}
\alias{enricher}
\title{Enrich pathways.}
\usage{
enricher(
  entrez_ids,
  species = "Mus.musculus",
  universe = NULL,
  showCategory = 20,
  pCutoff = 0.05,
  minGSSize = 10,
  qvalueCutoff = 0.2,
  pAdjustMethod = "none",
  filename_prefix = "enricher",
  output_folder = "."
)
}
\arguments{
\item{entrez_ids}{string vector of gene entrez id.}

\item{species}{Homo.sapiens or Mus.musculus, default is Mus.musculus.}

\item{universe}{same with enrichGO}

\item{showCategory}{same with dotplot}

\item{pCutoff}{adjusted pvalue cutoff on enrichment tests to report}

\item{minGSSize}{minimal size of genes annotated by Ontology term for testing.}

\item{qvalueCutoff}{qvalue cutoff on enrichment tests to report as significant. Tests must pass i) pvalueCutoff on unadjusted pvalues, ii) pvalueCutoff on adjusted pvalues and iii) qvalueCutoff on qvalues to be reported.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{filename_prefix}{a prefix of output filename.}

\item{output_folder}{which directory do you want to save to.}
}
\description{
Enrich pathways.
}
