% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts2fpkm.R
\name{fpkm2tpm}
\alias{fpkm2tpm}
\title{Convert FPKM to TPM}
\usage{
fpkm2tpm(fpkmExp)
}
\arguments{
\item{fpkmExp}{A data frame with FPKM values (rows are genes and columns are samples)}
}
\value{
a data frame with TPM values
}
\description{
Convert FPKM to TPM
}
