% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{get_sample_ids.Metadata}
\alias{get_sample_ids.Metadata}
\title{Get sample ids}
\usage{
\method{get_sample_ids}{Metadata}(m, subset_value = NULL)
}
\arguments{
\item{metadata}{a metadata object.}
}
\value{
a vector
}
\description{
Get sample ids
}
