% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{get_subset.Metadata}
\alias{get_subset.Metadata}
\title{Get a subset of metadata}
\usage{
\method{get_subset}{Metadata}(m, subset_value = NULL)
}
\arguments{
\item{metadata}{a metadata object.}

\item{subset}{which subset value in subset column.}
}
\value{
a data frame
}
\description{
Get a subset of metadata
}
