% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{logFC}
\alias{logFC}
\title{Compute logFC value.}
\usage{
logFC(group1, group2, is_log = TRUE)
}
\arguments{
\item{group1}{a vector which contains several values (default is log transformed).}

\item{group2}{a vector which contains several values (default is log transformed).}

\item{is_log}{boolean, default is TRUE. if your data is not log transformed, you need to set is_log to FALSE.}
}
\value{
a logFC value
}
\description{
Compute logFC value.
}
