% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes1pathway.R
\name{make.pathwaytable}
\alias{make.pathwaytable}
\title{Make a pathway table}
\usage{
make.pathwaytable(species.KEGG = "mmu")
}
\arguments{
\item{species.KEGG}{three-letter KEGG species identifier, only support mmu and hsa.}
}
\description{
Make a pathway table
}
