% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{make_groups.Metadata}
\alias{make_groups.Metadata}
\title{Make groups based on metadata table}
\usage{
\method{make_groups}{Metadata}(m, subset_value, group_column = "group", levels = c("PM", "FA"))
}
\arguments{
\item{group_column}{a field name.}

\item{metadata}{a data frame contains several columns for metadata.}
}
\value{
a factor
}
\description{
Make groups based on metadata table
}
\examples{
# Make groups
m <- data.frame(
  group = rep(c("FA", "PM"), 6),
  month = rep(5, 12), gender = rep("male", 12),
  organ = rep(c("gut", "lng", "lvr"), 4)
)
make_groups(m, group_column = "group", levels = c("PM", "FA"))
}
