% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{make_groups}
\alias{make_groups}
\title{Make groups based on metadata table}
\usage{
make_groups(metadata, group_column = "group", levels = c("PM", "FA"))
}
\arguments{
\item{metadata}{a data frame contains several columns for metadata.}

\item{group.column}{a field name.}
}
\value{
a factor
}
\description{
Make groups based on metadata table
}
\examples{
# Make groups
m <- data.frame(
  group = rep(c("FA", "PM"), 6),
  month = rep(5, 12), gender = rep("male", 12),
  organ = rep(c("gut", "lng", "lvr"), 4)
)
make_groups(m, group_column = "group", levels = c("PM", "FA"))
}
