% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes1pathway.R
\name{match.genelists}
\alias{match.genelists}
\title{Get the genes in a pathway.}
\usage{
match.genelists(pathways)
}
\arguments{
\item{pathways}{a data frame which contains two columns pathway_id and pathway_name}
}
\description{
Get the genes in a pathway.
}
