% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{Metadata}
\alias{Metadata}
\title{Metadata S3 class}
\usage{
Metadata(metadata_file)
}
\arguments{
\item{metadata_file}{a csv/tsv file}
}
\value{
a s3 class
}
\description{
Metadata S3 class
}
\examples{
# List all example files
system.file("extdata", package = "rOmicsUtils") |> list.files()

metadata <- rOmicsUtils::Metadata(file.path(system.file("extdata", package = "rOmicsUtils")[1], "GSE146508.csv"))
select_samples(metadata, subset_fields = c("organ", "name"))
make_groups(metadata, group_column = "group_name", subset_value = "Bld_Mutlu")
get_subset(metadata, subset = "Bld_Mutlu")

select_samples(metadata, sql_query_str = "organ = 'Bld' AND name = 'Biswal'", subset_fields = c("organ", "name"))
}
