% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{query_db4boxplot}
\alias{query_db4boxplot}
\title{Query DuckDB database}
\usage{
query_db4boxplot(db_path, query_str)
}
\arguments{
\item{db_path}{A file path for database file, such as my-db.duckdb}

\item{query_str}{A query string}
}
\value{
A data frame
}
\description{
Query parquet files with SQL query string.
}
