% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes1pathway.R
\name{read.pathways}
\alias{read.pathways}
\title{Read pathway file which contains two columns pathway_id and pathway_name.}
\usage{
read.pathways(file)
}
\arguments{
\item{file}{file path}
}
\description{
Read pathway file which contains two columns pathway_id and pathway_name.
}
