% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{remove_genes}
\alias{remove_genes}
\title{Remove low expression genes.}
\usage{
remove_genes(expr_mat, ratio = 0.6)
}
\arguments{
\item{expr_mat}{a matrix of data values.}

\item{ratio}{the ratio which the gene expression values equal 0 is greater than the inputed value, the related genes will be removed.}
}
\value{
a matrix which don't contain low expression genes.
}
\description{
Remove low expression genes.
}
