% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{save2pdf}
\alias{save2pdf}
\title{Write ggplot object to a file}
\usage{
save2pdf(plot, filename, filepath, width = 4, height = 4, units = "in")
}
\arguments{
\item{plot}{ggplot object.}

\item{filename}{a file name.}

\item{filepath}{a directory.}

\item{width}{image width.}

\item{height}{image height.}

\item{units}{Plot size in units ("in", "cm", "mm", or "px"). If not supplied, uses the size of current graphics device.}
}
\description{
Write ggplot object to a file
}
