% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{select_samples.Metadata}
\alias{select_samples.Metadata}
\title{Select samples}
\usage{
\method{select_samples}{Metadata}(m, subset_fields, sql_query_str = NULL)
}
\arguments{
\item{subset_fields}{which fields do you want to filter.}

\item{sql_query_str}{a query string.}

\item{metadata}{a data frame.}
}
\value{
a data frame
}
\description{
Select samples
}
