% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{ttest}
\alias{ttest}
\title{Performs one and two sample t-tests on vectors of data.}
\usage{
ttest(gene, x, y, paired = FALSE)
}
\arguments{
\item{gene}{a gene symbol/id.}

\item{x}{a (non-empty) numeric vector of data values..}

\item{y}{an optional (non-empty) numeric vector of data values.}
}
\value{
a data frame which contains gene, P.Value, statistic, and logFC columns.
}
\description{
Performs one and two sample t-tests on vectors of data.
}
