% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{which_database}
\alias{which_database}
\title{Guess database which the gene ids come from.}
\usage{
which_database(genes)
}
\arguments{
\item{genes}{string vector of gene id.}
}
\value{
database name, ENSEMBL, ENTREZID or SYMBOL.
}
\description{
Guess database which the gene ids come from.
}
