% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{annot}
\alias{annot}
\title{Annotate Gene ID}
\usage{
annot(geneids, species = "Homo.sapiens")
}
\arguments{
\item{geneids}{Ensembl ids}

\item{species}{On support "Homo.sapiens" or "Mus.musculus"}
}
\value{
A data frame
}
\description{
Annotate ensembl id
}
